Fix remaining float to double conversions since b1ccea2c79dea8eee98c628a74adce3687df61ea

+++ src/main/java/evilcraft/events/LivingDeathEventHook.java
@@ -40,8 +40,8 @@ public class LivingDeathEventHook {
                 && event.entityLiving != null && Configs.isEnabled(BloodExtractorConfig.class)) {
             EntityPlayerMP player = (EntityPlayerMP) e;
             float health = event.entityLiving.getMaxHealth();
-            int minimumMB = MathHelper.floor_float(health * BloodExtractorConfig.minimumMobMultiplier);
-            int maximumMB = MathHelper.floor_float(health * BloodExtractorConfig.maximumMobMultiplier);
+            int minimumMB = MathHelper.floor_float(health * (float) BloodExtractorConfig.minimumMobMultiplier);
+            int maximumMB = MathHelper.floor_float(health * (float) BloodExtractorConfig.maximumMobMultiplier);
             BloodExtractor.getInstance().fillForAllBloodExtractors(player, minimumMB, maximumMB);
         }
     }
